+++ examples/basic/basic_language_model_t5_uer.py
@@ -7,7 +7,6 @@ from bert4torch.models import build_transformer_model
 from bert4torch.tokenizers import Tokenizer, load_vocab
 from bert4torch.generation import AutoRegressiveDecoder
 
-# bert配置
 config_path = 'E:/pretrain_ckpt/t5/[uer_t5_torch_small]--t5-small-chinese-cluecorpussmall/bert4torch_config.json'
 checkpoint_path = 'E:/pretrain_ckpt/t5/[uer_t5_torch_small]--t5-small-chinese-cluecorpussmall/pytorch_model.bin'
 dict_path = 'E:/pretrain_ckpt/t5/[uer_t5_torch_small]--t5-small-chinese-cluecorpussmall/vocab.txt'